v1
Small iteration code to reinforce how to properly utilise a loop to
iterate as it confuses me for some reason. Also made the code work so I
just put all testing code in it's own little function with all it's test
variables private to that function, then call on it in main() for
testing.